[refactor] 헤더 드롭다운 컴포넌트 분리
