+++ src/components/layout/Header.tsx
@@ -2,24 +2,16 @@
 
 import Link from 'next/link';
 import MainLogo from '@/assets/logo_horizontal_1.png';
-import { useEffect, useState } from 'react';
-import { toast } from 'react-toastify';
-import { useAuth } from '@/hooks/useAuth';
+import Image from 'next/image';
+import { useEffect } from 'react';
 import { useAtom } from 'jotai';
 import { AuthChangeEvent } from '@supabase/supabase-js';
-import { DropdownItem, DropdownTrigger, Dropdown, DropdownMenu, Avatar } from '@nextui-org/react';
 import { isLoggedInAtom, isMenuOpenAtom } from '../../store/store';
 import { supabase } from '@/utils/supabase/supabase';
-import { User } from '@/types/users';
-import { profileStorageUrl } from '@/utils/supabase/storage';
-import { useQuery } from '@tanstack/react-query';
-import { getUser } from '@/api/users';
-import Image from 'next/image';
+import ProfileDropdown from './ProfileDropdown';
 
 const Header = () => {
   const [isLoggedIn, setIsLoggedIn] = useAtom(isLoggedInAtom);
-  const [currentUser, setCurrentUser] = useState<User | null>(null);
-  const { logout, getCurrentUserProfile } = useAuth();
 
   /** 현재 로그인되어 있는지 확인 */
   useEffect(() => {
@@ -73,86 +65,29 @@ const Header = () => {
     };
   }, [setIsLoggedIn]);
 
-  /** 로그인한 사용자의 정보를 profiles 테이블에서 불러옴 */
-  const { data, isLoading, isError } = useQuery<User | null>({
-    queryKey: ['loggedInUser'],
-    queryFn: async () => {
-      try {
-        const getSession = await supabase.auth.getSession();
-        if (!getSession.data.session) return null;
-        const fetchData = await getUser(getSession.data.session.user.id);
-        // console.log('ㅇㅇ', fetchData);
-        setCurrentUser(fetchData);
-        return fetchData;
-      } catch (error) {
-        throw new Error('사용자 정보를 가져오는 데 실패했습니다.');
-      }
-    },
-    refetchOnWindowFocus: false
-  });
-
-  if (isLoading) return <div>로그인 정보를 불러오고 있습니다.</div>;
-  if (isError) return <div>데이터 로드 실패</div>;
-
-  /** 로그아웃 핸들러 */
-  const handleLogout = async () => {
-    await logout();
-    setIsLoggedIn(false);
-    toast.success('로그아웃되었습니다.');
-  };
-
   return (
-    <>
-      <header className="w-full h-[8vh] leading-[7.5vh] flex text-pointColor1 font-bold bg-bgColor1 border-solid border-b-2 border-pointColor1">
-        <Link href="/" className="w-[16%] text-center flex justify-center items-center">
-          <Image src={MainLogo} alt="로고" width={150} />
-        </Link>
-        <section className="w-[84%] flex justify-between px-10">
-          <nav className="flex gap-14">
-            <Link href="/quiz-list" className="">
-              퀴즈
-            </Link>
-            <Link href="/typing-game">타자 연습</Link>
-            <Link href="/community-list?category=전체">커뮤니티</Link>
-            <Link href="/about">서비스 소개</Link>
-          </nav>
-          {isLoggedIn ? (
-            <div className="flex items-center gap-4">
-              <Dropdown placement="bottom-end">
-                <DropdownTrigger>
-                  <Avatar
-                    as="button"
-                    className="transition-transform"
-                    color="default"
-                    size="md"
-                    src={`${profileStorageUrl}/${data?.avatar_img_url}`}
-                  />
-                </DropdownTrigger>
-                <DropdownMenu aria-label="Profile Actions" variant="flat">
-                  <DropdownItem key="profile" className="h-14 gap-2">
-                    <p className="font-semibold">Signed in as</p>
-                    <p className="font-semibold">{currentUser?.email}</p>
-                  </DropdownItem>
-                  <DropdownItem as={Link} href="/profile">
-                    내 프로필
-                  </DropdownItem>
-                  <DropdownItem as={Link} href="/my-activity">
-                    나의 활동
-                  </DropdownItem>
-                  <DropdownItem key="logout" color="danger" onClick={handleLogout}>
-                    Log Out
-                  </DropdownItem>
-                </DropdownMenu>
-              </Dropdown>
-            </div>
-          ) : (
-            <Link href="/login">
-              <button>로그인</button>
-            </Link>
-          )}
-        </section>
-      </header>
-    </>
+    <header className="w-full h-[8vh] leading-[7.5vh] flex text-pointColor1 font-bold bg-bgColor1 border-solid border-b-2 border-pointColor1">
+      <Link href="/" className="w-[16%] text-center flex justify-center items-center">
+        <Image src={MainLogo} alt="로고" width={150} />
+      </Link>
+      <section className="w-[84%] flex justify-between px-10">
+        <nav className="flex gap-14">
+          <Link href="/quiz-list" className="">
+            퀴즈
+          </Link>
+          <Link href="/typing-game">타자 연습</Link>
+          <Link href="/community-list?category=전체">커뮤니티</Link>
+          <Link href="/about">서비스 소개</Link>
+        </nav>
+        {isLoggedIn ? (
+          <ProfileDropdown />
+        ) : (
+          <Link href="/login">
+            <button>로그인</button>
+          </Link>
+        )}
+      </section>
+    </header>
   );
 };
